--- src/main/java/dgu/umc_app/global/service/TokenBlacklistService.java
@@ -1,56 +1,0 @@
-package dgu.umc_app.global.service;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.stereotype.Service;
-
-import java.time.Duration;
-
-@Slf4j
-@Service
-@RequiredArgsConstructor
-public class TokenBlacklistService {
-
-    private final RedisTemplate<String, String> redisTemplate;
-    private static final String BLACKLIST_PREFIX = "blacklist:";
-    private static final String REFRESH_TOKEN_PREFIX = "refresh_token:";
-
-    // 액세스 토큰을 블랙리스트에 추가
-    public void addToBlacklist(String token, long expirationTimeInSeconds) {
-        String key = BLACKLIST_PREFIX + token;
-        redisTemplate.opsForValue().set(key, "blacklisted", Duration.ofSeconds(expirationTimeInSeconds));
-        log.info("토큰이 블랙리스트에 추가됨: {}", token);
-    }
-
-    // 리프레시 토큰을 저장
-    public void saveRefreshToken(String userId, String refreshToken, long expirationTimeInSeconds) {
-        String key = REFRESH_TOKEN_PREFIX + userId;
-        redisTemplate.opsForValue().set(key, refreshToken, Duration.ofSeconds(expirationTimeInSeconds));
-        log.info("리프레시 토큰 저장됨: userId={}", userId);
-    }
-
-    // 토큰이 블랙리스트에 있는지 확인
-    public boolean isBlacklisted(String token) {
-        String key = BLACKLIST_PREFIX + token;
-        return Boolean.TRUE.equals(redisTemplate.hasKey(key));
-    }
-
-    // 리프레시 토큰 조회
-    public String getRefreshToken(String userId) {
-        String key = REFRESH_TOKEN_PREFIX + userId;
-        return redisTemplate.opsForValue().get(key);
-    }
-
-    // 리프레시 토큰 삭제
-    public void deleteRefreshToken(String userId) {
-        String key = REFRESH_TOKEN_PREFIX + userId;
-        redisTemplate.delete(key);
-        log.info("리프레시 토큰 삭제됨: userId={}", userId);
-    }
-
-    // 사용자의 모든 토큰 삭제 (로그아웃)
-    public void logoutUser(String userId) {
-        deleteRefreshToken(userId);
-    }
-}
